Refactor: move combinations function to d2.py and update d3.py to use it
