+++ d3.py
@@ -1,4 +1,4 @@
-from f2 import combinations
+from d2 import combinations
 
 n = 5
 r = 2
